Move IDM up in CODEOWNERS (#6062)
## Summary of changes

This will allow all of the others to override IDM more consistently.

## Reason for change

IDM was getting added to all of the CI Visibility PRs and this should
address that.

## Implementation details

- Select lines
- `Ctrl + X`
- `Ctrl + V`
- Select lines
- `Ctrl + X`
- `Ctrl + V`
- `Ctrl + S`
- `Ctrl + S`
- `Ctrl + S`

## Test coverage

:shrug:

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->

---------

Co-authored-by: Tony Redondo <tony.redondo@datadoghq.com>