[ASM] Fix bug RC empty key (#6058)
## Summary of changes

Fixed a regression bug.
The key variable for the process update wasn't the right one.

# Bug explanation

The Path of the file wasn't get from the right variable.
`result.TypedFile` is a `JToken`, getting the `Path` from this object
would get the path of the JSON token. The path is an empty string and
totally not the info we wanted to get.

## Test coverage

- Add a way to set custom headers instead of the default `Mistake
Not...` to trigger a blocking event from the default ruleset
- Added an integration test where it tests deletion and switch back to
default rules but keep updated blocking action.