[IAST] Improve RestSharp SSRF detection (#6060)
## Summary of changes
Improvements in the tainting of string.Replace and string.Format.
Improvements in escape functions.
Addition of SSRF secure mark.
Instrumentation of UrlEscape methods.

## Reason for change
Inaccurate detection of SSRF in RestSharp queries have been reported.

## Implementation details
Added a check to verify if the result string is the same as the original
when tainting the whole string.
This happens in Replace and Format strings, when the function does not
modify the input.
For instance `"text".Replace("1", "2")` will return `"text"`, but if
`"2"` is tainted, the result would be also, and this was incorrect.

Fixed the escape function to propagate the secure mark in the escaped
function, and not in the origin, as was done until now.

Added a `CallTarget` in the `UrlEncode` to properly escape the result
and propagate a secure mark for SSRF, for when `AddParameter` is called
with the `encode` argument to true (by default).

## Test coverage
Added unit tests to check the new proper `Replace `and Format
functionality, as well as the `UrlEncode `and the `RestSharp `SSRF
detection

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->