Add a outer startup hook catch block (#6068)
## Summary of changes

Before, the `__DDVoidMethodCall__` method loaded the loader and had a
try/catch block.

Now `__DDVoidMethodCall__` still has a try/catch block, but it calls
`__DDInvokeLoader__` which loads the loader.

## Reason for change

When running in a trimmed app, the JIT compilation of
`__DDVoidMethodCall__` would throw an exception. Moving the logic to an
inner method allows this exception to be caught.

## Implementation details

Just added a method and moved the catch block.

## Test coverage
Running yuniql crashes without this change.