Add `ddprs` to DBM injected comment + fix `dddbs` (#6084)
## Summary of changes

- add `ddprs` (peer service) to the comment injected by DBM, contains
the peer service value only if set explicitely
- revert the behavior introduced in #4619, as per spec, peer service is
not supposed to override `dddbs`.

## Reason for change

spec:
https://docs.google.com/document/d/1Cg9RLydNzK1BTf3EFaEC4NlQtvAO-_rDdPLSO8HgDCs
slack thread:
https://dd.slack.com/archives/C02U25426A1/p1727254455343209

## Implementation details

By taking the value of `ServiceNameInternal` from the span, dddbs will
have either the service name of the calling service or the name of the
DB depending on how fake services are configured.

The SQL tags V1 record if a peer service override was set

https://github.com/DataDog/dd-trace-dotnet/blob/40c8f59154c71dcce502c740304b9ef281f5c355/tracer/src/Datadog.Trace/Tagging/SqlTags.cs#L57-L58
so I'm using that to know if I should use the `PeerService` value
(because it always has a value, it falls back to the DB type or the
outhost if peer service is not set).

https://github.com/DataDog/dd-trace-dotnet/blob/40c8f59154c71dcce502c740304b9ef281f5c355/tracer/src/Datadog.Trace/Tagging/SqlTags.cs#L48

For V0, I'm relying on the fact that the old value is saved when peer
service is set. Not sure if I could just always use peer service in that
case ?

https://github.com/DataDog/dd-trace-dotnet/blob/4bafc78fa25382fefb03fa4e052905ca371d1ffa/tracer/src/Datadog.Trace/Configuration/Schema/NamingSchema.cs#L64-L65

## Test coverage

## Other details
JIRA: AIT-10032
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->