Minor build fixes (#6099)
## Summary of changes

- Ensure we always upload logs, even if a job is cancelled
- Don't clean test results multiple times per run

## Reason for change

We have some cases where integration tests are taking _way_ too long to
run, and CI times out and cancels the job. Currently, that means we also
don't get any logs.

Also, we're (accidentally) cleaning the test log folder multiple times
in some stages. For example, we run Windows integration tests and
regressions tests (two different stages) in the CI with a single
command. Both of the run steps will clear the test folder, so we end up
losing the integration test logs (unless they fail, in which case the
regression tests never run)

## Implementation details

- Changed from `succeededOrFailed()` to `always()` for the log and
snapshot upload steps
- Extracted a "clean test folder" target which runs automatically _once_
for the full execution.

## Test coverage

This is the test