Run CallTargetNativeTests on Windows (#5754)
## Summary of changes

Runs the `CallTargetNativeTests` on Windows

## Reason for change

We weren't running these on Windows, and we should be

## Implementation details

Add the `RunOnWindows` trait. Longer term I want to reverse the role of
this trait, i.e. we run on windows by default and you add
`RunOnWindows=False` to disable that.

## Test coverage

More now

## Other details

~~I thought we had a bug related to ref structs here, but it appears to
be a Schrödinger's bug~~ Oops, no, there it is!