[ASM] Suspicious attacker blocking (#6057)
## Summary of changes

This PR adds the required code to support the suspicious attack
functionality.

It has added the required code to read the exclusion data from the RC.

It also has modified the existing code regarding RC actions. Previously,
if a configuration was received with an action, the configuration was
stored and later sent to the WAF. If new values with an empty action
array would come later, the previous action configurations would be
deleted, but if a new array would come with a new action different than
the previous one, we would report them both to the WAF. This behavior
was making the suspicious attacker system tests fail because we would
keep RC changes from previous tests. This change seems to be aligned
with the behavior of other libraries.

The file
[AspNetBase.cs](https://github.com/DataDog/dd-trace-dotnet/pull/6057/files#diff-0faff2451113067d7669566ba9199908b720a3764914b00d6f33d4b376098d74)
has been updated. Now, tests have more control over the used headers by
allowing them to remove headers or replace previous values with new
ones.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->