[Crashtracking] Implement support for Windows (#6088)
## Summary of changes

Add crashtracking support for Windows.

At the moment, only x64 is supported. Supporting x86 would require to
include a 32-bit version of `dd-dotnet` in the package.

## Reason for change

Until now, crashtracking only supported Linux. There are many crashes
(due to either netfx or IIS) that only occur on Windows.

## Implementation details

Crashtracking on Windows relies on
[WER](https://learn.microsoft.com/en-us/windows/win32/wer/windows-error-reporting).

When the native loader is loaded, we use
`WerRegisterRuntimeExceptionModule` to register it as a crash handler.
Note that .NET also registers a crash handler, and they're invoked by
order of registration. To be invoked first, we unregister the .NET crash
handler, register our own, then re-register the .NET crash handler.

The .NET crash handler is in the DAC (mscordawks.dll for .NET Framework,
mscordaccore.dll for .NET Core). The context argument is set to the
address of `clr.dll`/`coreclr.dll` in memory (this is important because
`WerUnregisterRuntimeExceptionModule` fails if the context value is
different).

When a crash occurs, Windows freezes the process and launches
`WerFault.exe`. In WerFault, the crash handler is loaded and given a
chance to inspect the crash (using the exported
`OutOfProcessExceptionEventCallback` function). We invoke `dd-dotnet` to
inspect the process and generate the crash report.

The environment variables of the crashing process are not propagated to
`WerFault.exe`. Because we need them (for instance, to know the location
of the agent), we make a copy of all the `DD_*` environment variables
and give the address of that copy as the "context" argument of
`WerRegisterRuntimeExceptionModule`. When
`OutOfProcessExceptionEventCallback` is invoked in `WerFault.exe`, we
use `ReadProcessMemory` to read the saved environment variables from the
crashing process, and set them for `dd-dotnet`.

## Test coverage

Reusing the existing crashtracking tests.

## Other details

More information on WER: https://minidump.net/windows-error-reporting/