[Crashtracking] Add registry key to installer (#6106)
## Summary of changes

Register `Datadog.Trace.ClrProfiler.Native.dll` in the
`HKLM\Software\Microsoft\Windows\Windows Error
Reporting\RuntimeExceptionHelperModules` key.

## Reason for change

This is needed by crashtracking.
If the value is absent, the tracer will create it in HKCU (the
instrumented application is unlikely to have permission to write to
HKLM). However, registering the crash handler in HKCU is only supported
in somewhat versions of Windows (since Windows 10 2004). So we should
rely on HKLM whenever possible.

## Other details

Depends on https://github.com/DataDog/dd-trace-dotnet/pull/6088

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>