[IAST] Avoid lock in method ctor. (#6115)
## Summary of changes
Avoid an unneeded lock in the `MethodInfo` constructor when calculating
full name

## Reason for change
All methods need to calculate their full name in order to determine if
they are excluded from dataflow. FullName calculation is lazy, and thus,
protected by a mutex. This was moved to the constructor in a previous
refactor, but the mutext kept bein taken, and there's no need for that

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->