Verify and assert version of glibc targeted in native tracer/profiler (#6104)
## Summary of changes

- ~Bail out of tracing if glibc < 2.17 on x64 or < 2.23 on arm64~
- Add validation checks to the native libraries to ensure we don't
exceed these

## Reason for change

The native tracer and profiler are compiled against glibc/musl and as
such have minimum versions they require - the musl version is
_generally_ not an issue, but glibc can be. The required glibc version
is _generally_ controlled by the system on which it's built. For our
native build we have

- x64 - CentOS 7 with glibc `2.17`
- arm64 - Debian 10 with glibc `2.28`

If we run on a glibc version that is too low, then we _don't_ fail to
load because we use `RTLD_LAZY` lazy binding (instead of `RTLD_NOW`):

> Perform lazy binding. Resolve symbols only as the code that references
them is executed. If the symbol is never referenced, then it is never
resolved. (Lazy
binding is performed only for function references; references to
variables are always immediately bound when the shared object is
loaded.) Since glibc 2.1.1,
this flag is overridden by the effect of the LD_BIND_NOW environment
variable.

Ultimately, this means that we might load but crash _later_ when a
symbol that is not available is required.

To guard against that, we _could_ add checks for the glibc version into
the native loader, and bail out if we detect a low glibc version.

Unfortunately, things get tricky with arm64, where we _technically_
require `2.28`. .NET 5-8 only require 2.23, and the popular [Amazon
Linux 2 distro only requires
2.26](https://docs.aws.amazon.com/linux/al2023/ug/glibc-gcc-and-binutils.html).
So if we were to block <2.28 that could be a significant problem.

The reason this isn't a problem _today_ is that while the profiler
requires `2.28`, it's not currently supported on arm64. The native
tracer, meanwhile, currently only requires `2.18`.

The longer term solution is to fix our min glibc dependency to match
.NET's `2.23` requirement, but that requires a significant investment,
likely using a [rootfs build
technique](https://github.com/dotnet/runtime/blob/main/docs/workflow/building/coreclr/cross-building.md#linux-cross-building)
like [the .NET runtime
does](https://github.com/dotnet/runtime/issues/83428) or something like
https://github.com/wheybags/glibc_version_header.

In the meantime, we will bail out on the following:

- x64 - glibc < `2.17`
- arm64 - glibc < `2.23`. Note _not_ `2.28`

This will be fine as long as 

- The native tracer doesn't require symbols from > `2.23`.
- The continuous profiler doesn't enable arm64.

~This PR enforces the former condition by verifying the required glibc
version in the native binary. We cannot enforce the latter currently, so
will need to bear this in mind when we enable arm64 profiling.~ Given
that those thresholds are also the minimum for .NET, we will _not_
verify them in the native loader directly, instead we will rely on tests
to assert we don't exceed 2.23 in the native tracer.

## Implementation details

~The glibc check in the native loader uses the same technique as the SSI
injector in https://github.com/DataDog/auto_inject/pull/433 (which is
compatible with musl, and even busybox) to retrieve the glibc version.~

To retrieve the actual glibc symbols used by the native library, it's
effectively running the following command:

`nm --with-symbol-versions -D <TRACERFILE> | grep GLIBC | sed
's/^.*GLIBC_//' | sort -u -t "." -k1,1nr -k2,2nr -k3,3nr | head -1`

Which is obviously easier to follow in C# 😉 The `nm` command returns a
list of symbols, similar to this:

```
U __iswctype_l@GLIBC_2.17
U __lxstat@GLIBC_2.17
U __newlocale@GLIBC_2.17
```

We just grab the suffix and find the highest: that's our _real_ minimum
glibc version.

One complication is that the version of `nm` in centos7 is too old, so
it doesn't work 🙄 So we have to perform the check on the debian distro.

## Test coverage

I've tested this all manually, and I'll run against the full smoke tests
to be sure. But it's kind of hard to test in practice, given we're
basically limiting our self to .NET's glibc version now.

## Other details

Given that final statement, a reasonable question is "why bother then"?
Which is reasonable - I started down this route precisely because we
_actually_ require `2.28`. So we can always park this if we want.

EDIT: Yeah, decided to drop the check in the native loader in the end