[backport] [Tracer] Set _dd.base_service tag whenever a span's service name is different than the default value (DD_SERVICE) (#5502 => v2) (#6122)
## Summary of changes

This fixes the setting of the `ServiceName` so that instead of just
adding a `_dd.base_service` tag when `Span.ServiceName` is overridden
that it'll be computed during MessagePack serialization and checked to
see if it is different than the default value (aka `DD_SERVICE`).

If it is different from the default then we add the default name as
`_dd.base_service`

## Reason for change

We want to record the base service name after the service name has been
updated.

## Implementation details

- Remove the setting of this in `Span.ServiceName`
- Add check in `SpanMessagePackFormatter` for that.

## Test coverage

- Updated snapshots

## Other details
Backport of #5502

<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->