[IAST] Calculate method full name on demand always (#6127)
## Summary of changes
Stop caching method's fullname in MethodInfo and calculate it on demand

## Reason for change
FullName caching could led to race conditions, and uses memory. It is
only used on method creation for exclusion decision, on debug traces and
when instrumentation fails, so there's no need for caching it, freeing
up memory in the process.

## Implementation details
Remove `_fullName` member in `MethodReference `class and calculate it
always when calling `GetFullName()`

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->