[ASM] Update fingerprint rules (#6133)
## Summary of changes

This PR updates the format of the fingerprint rules as described in the
[RFC](https://docs.google.com/document/d/1DivOa9XsCggmZVzMI57vyxH2_EBJ0-qqIkRHm_sEvSs/edit?pli=1#heading=h.88xvn2cvs9dt)

We don't send anymore the { "fingerprint", true } request to the WAF.
Instead, the WAF returns the fingerprints when a match occurs. This way,
we save one WAF call.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->