[ASM] Fix one flakiness in rcm asm data integration tests and simplify some asm rcm code (#6119)
## Summary of changes

Problem was: we send asm features.asm.enabled=true at the same time as
we sent some asm_data. But, we need to rememeber that when asm is off,
there is NO asm_data, asm products rc subscriptions. So we're not
expecting these products hence a common warning in the logs like:
> Received config RemoteConfigurationPath { Path =
datadog/2/ASM_DATA/AspNetCore5AsmDataSecurityEnabledBlockingRequestIpOneClick8f9d4f72-ea07-478c-ae2f-4cf25117a1e7/config

It was flaky and still working because as soon as
asm_features.asm.enabled = true is applied, we apply other
configurations that continue to be sent right after (as the mock agent
is setup with these files and will give the same response if unchanged).
So most of the time, the 2 rc sets would be applied quickly and the
request sent right after would show a security event.

But, we weren't waiting for the right state, as we were waiting for only
one rc response (with asm_features _and_ asm_data together) and we know
they can't be applied at the same time (as the asm_data subscription is
removed). It's applied right after. That's why, most of the time, it
worked. But it shouldn't really work as we should have been waiting for
another round of rc polling. So now, we separate both requests.

So conclusion: beware of what subtle flake can turning on/off
subscriptions cause.
We still want to keep behaving the same though, we dont want to have
useless products subscriptions if appsec is disabled, even if that means
a creation then later an update of the waf..

## Reason for change

Simplify some code as well, no need for the fallback to embedded rule
set flag, just test if "rules" is in the dictionary, if not add it from
the embedded set.
## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->