[ASM] Send header values as string to the WAF (#6144)
## Summary of changes

Util this PR, we woudl store the request headers in a array of strings
and we would send that array to the WAF. When generating fingerprints,
the WAF does not accept arrays of strings but a single string, so, when
we have a single header value, we store it in a string instead of in an
array of strings with a single value inside the array.

This change has fixed some problems related to the fingerprint
generation, which has derived in a large amount of updated snapshots.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->