Handle Nuke .NET 9 preview issue (#6130)
## Summary of changes

- Adds workaround for case where .NET 9 (preview) SDK is installed
lately.

## Reason for change

Fixes case where Nuke no longer works if you install the .NET 9  SDK

## Implementation details

Currently we have a global.json which enforces that we use the .NET
8.0.100 SDK for building _in general_. However, we have to use the .NET
7 SDK for running the Nuke project specifically when we're building the
Linux x64 _native_ components.

> This is because .NET 8 _doesn't_ support CentOS 7 and can't run there,
but we _have_ to use CentOS 7 currently for technical reasons.

We worked around this previously by specifically opting in to
major-version rollforward in the Nuke project. Unfortunately, Nuke
doesn't work out of the box with .NET 9 because of the [binary

formatter](https://learn.microsoft.com/en-gb/dotnet/standard/serialization/binaryformatter-security-guide)
removal. Consequently we have to make sure we _don't_ roll forward to
.NET 9 in that case.

The solution is pretty simple:
- Explicitly use .NET 8 _without_ roll forward in general
- Set an env var in the CentOS7 dockerfile `USE_NATIVE_SDK_VERSION=true`
which switches the Nuke project to build and run on .NET 7 explicitly

## Test coverage

Tested locally that it fixes the .NET 9 issue, this is the test for CI

## Other details

~The side-quest here was updating Nuke as it was quiet out of date. I
_thought_ it might be required, and it required fixing a _bunch_ of
warnings and breaking changes, so figured it would make sense to just do
the work here at the same time~ This ended up being a horrible pit of
errors so I abandoned it as it seems to be unnecessary.