Stop using current_path in the native loader (#6132)
## Summary of changes

Stop changing the current path in the native loader.

## Reason for change

We've received crash reports caused by `fs::current_path`. It's unclear
why it threw (apparently it can happen for instance if the current path
is too long), but it doesn't look like we actually need it anyway (it
was needed for `fs::absolute` but we can replace it by directly
appending the relative path to the base path).

## Implementation details

After removing the call to `current_path`, there is a risk that whatever
caused it to fail would cause `fs::exist` to fail as well, so I replaced
the call with the overload that doesn't throw.

## Test coverage

Many tests already rely on the native loader.