Subscribe to AssemblyLoadContext.Default.Resolving (#6148)
## Summary of changes

In the managed loader, subscribe to
`AssemblyLoadContext.Default.Resolving` in addition to
`AppDomain.CurrentDomain.AssemblyResolve`.

## Reason for change

When an assembly is loaded implicitly (because it's referenced in some
code), the JIT compiler calls
`AssemblyLoadContext.ResolveUsingResolvingEvent`. This method asks the
current ALC to resolve the assembly, and it will in turn raise its
`Resolving` event if it fails. If the active ALC `Resolving` event
fails, then the global `AssemblyLoadContext.AssemblyResolve` event is
finally raised.

In the managed loader, we subscribe to
`AppDomain.Current.AssemblyResolve`, which is a wrapper around
`AssemblyLoadContext.AssemblyResolve`. It means that the ALC `Resolving`
event has a chance to resolve assemblies before we do.

`System.Management.Automation`, a powershell library, subscribes to
`AssemblyLoadContext.Default.Resolving`. In its resolution logic, it
loads assemblies from the GAC if they're not found, thus loading the
.NET Framework version of Datadog.Trace even if the process is running
.NET Core. By subscribing to `AssemblyLoadContext.Default.Resolving`, we
prevent that from happining.

## Test coverage

Added a smoke test.

## Other details

Fixes https://github.com/DataDog/dd-trace-dotnet/issues/6135