[ASM] Rename snapshots for ASM ownership (#6155)
## Summary of changes

We have two snapshots that don't follow the naming conventions. This
means that approval from other teams will be required when there is a
change in either of them. Renaming them to follow the conventions will
prevent this situation.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->