Fix return value on error in delegate instrumentation (#6153)
## Summary of changes

In the delegate instrumentation, when the async callback failed, we
returned a default value instead of the original value.

## Reason for change

The instrumented method probably had a good reason to return that value
to begin with.

## Implementation details

Just set the value in the catch block.

## Test coverage

Added tests to cover the "failure in callback" case.

Also, fixed a flake in the async test (which is actually the reason why
I looked at this code to begin with). The tests were relying on the
order of execution of the `onDelegateEnd`/`onDelegateAsyncEnd`
callbacks, but their order is not guaranteed (because
`onDelegateAsyncEnd` is invoked asynchronously and not awaited).