[CI Visibility] Ensure commit messages are trimmed (#6170)
## Summary of changes

In this PR we ensure that the commit message gets trimmed.

## Reason for change

We saw this issue in v2.41.0, I couldn't reproduce it in 3.x but let's
do this just to be sure.

## Implementation details

Call string.Trim

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->