[ASM] Remove httpcontext from context store at the end of the request (#6151)
## Summary of changes

Some logs show that we're accessing HttpContext properties when it has
already been disposed. Stacktrace shows that it's at the beginning of
running the waf, so this might be a simple fix for now

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->