[ASM] Send cookie values as single string to the WAF (#6164)
## Summary of changes

This PR sends the request cookies to the WAF as a single string when
there is only one cookie value. Previously, we would send the value in a
list. This makes the WAF to ignore these cookies when calculating the
session fingerprint.

The code for header and cookies collection has been moved to
securitycoordinator.cs in order to avoid duplicated code. It also helps
to have a consistent behavior between frameworks. For instance, when we
had no cookies, we would send an empty dictionary to the WAF in .net
framework and we would send nothing for other frameworks.

As a result, the fingerprint generations have changed. A new cookie has
been added to some integration tests to make sure that the hash is
calculated.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->