[ASM] Add rules version to all spans when ASM enabled (#6188)
## Summary of changes

This PR adds the _dd.appsec.event_rules.version tag to all the spans
when ASM is enabled. This has been requested by the RASP WG and is
consistent with the behavior of other libraries.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->