[Profiler] Make sure we log only once for DebugInfoStore errors (#6187)
## Summary of changes

Make sure that in case of errors, the DebugInfoStore (line number
feature) logs only once the error per module.
## Reason for change
When enabling the "line number" feature, if a module is invalid (does
not exist, discrepancies between the pdb and the image...), we log an
error each time we ask for debug info.
```
[2024-10-23 12:46:18.933 | info | PId: 21 | TId: 21] The debug info for the module `140011354144768` seems to be invalid (is valid? '0') or the RID is out of the symbols array bounds (RID: 23527). Number of debug info: 0
```
This clutters the log file and does not help.

## Implementation details

Add a field per module datatstructure that says if the error was already
logged for that module.

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->