Inject startup hook preferentially into static constructor when available (#6154)
## Summary of changes

Before injecting the startup hook, check if the type has an explicit
static constructor. If so, skip injecting.

## Reason for change

This fixes an issue with manual instrumentation in v3 when the
entrypoint contains an explicit static constructor. A simple
reproduction is the following:

```csharp
public static class Program
{
    private static readonly Datadog.Trace.Tracer _tracer = Datadog.Trace.Tracer.Instance;
    static Program()
    {
        _ = _tracer;
    }

    public static void Main(string[] args)
    {
        var builder = WebApplication.CreateBuilder(args);
        var app = builder.Build();

        app.MapGet("/", () => {
            using var scope = _tracer.StartActive("custom-operation-name");
            return "Hello World!";
        });

        app.Run();
    }
}
```

In this example, we inject into the entrypoint `Program.Main()`. We add
the startup hook which initializes the instrumentation. However, the JIT
sees that there is a static constructor and manually inserts a call to a
static field to trigger its execution. This is inserted at the start of
`Program.Main` i.e. _before_ our startup hook.

> Note that this _only_ happens in .NET Core - in .NET Framework the
`.cctor` is re-jitted before the entrypoint, so the problem doesn't
exist. So we limit this implementation to .NET Core to reduce the blast
radius.

In this situation, we don't see the initialization of the tracer, so we
miss the ReJit calls, and the manual instrumentation library ends up in
a half-instrumented state.

This is particularly noticeable for the manual instrumentation case,
because we have a lot of instrumentation, but technically it applies to
anything that is called from a static constructor in the entrypoint.


## Implementation details

When choosing whether to inject the startup hook, check whether the
entrypoint type contains a static constructor. If it does, _and the
constructor is not an implicit constructor_, then skip instrumenting, on
the basis that the JIT will subsequently inject a call to the static
constructor, which we will then instrument, ensuring we _are_ the first
call in the app.

In some cases (i.e. IIS) we're specifically choosing where we want to
instrument
(`System.Web.Compilation.BuildManager.InvokePreStartInitMethods()`) so
we don't try to change anything in that case.

The implicit static constructor is annoying. If there's a static field,
but no explicit static constructor, then the profiling API will return a
static constructor. However, we _shouldn't_ skip on the basis of this
implicit constructor, because `JITCompilationStarted` is never called
for it, so we would end up initializing too late. To detect this, [we
check whether `beforefieldinit` is
set](https://csharpindepth.com/Articles/BeforeFieldInit) to try to
decide whether to rely on the static constructor actually being called
or not.

There's obviously a risk with all this; causing things to initialize in
the wrong order is a perennial problem. If for some (strange) reason the
static constructor is _not_ invoked next, then we may end up injection
somewhere else, which would be... strange... but also Bad™😅 At least
we're not doing this behavior in .NET FX which is often the source of
weirdness!

## Test coverage

Added a reproduction to `Samples.ManualInstrumentation`. Without the
change to startup hook injection the tests fail at
`ThrowIf(string.IsNullOrEmpty(Tracer.Instance.DefaultServiceName))`
(because it _is_ empty) and also at `Tracer.Instance.StartActive()`

## Other details

Discovered this behaviour while trying to solve #6124
#6184 actually resolved the issue, so this is just an edge case